Moved the stage_local_data_in_gcs and upload_to_gcs functions to gcs_utils module
