+++ google/cloud/aiplatform/utils/__init__.py
@@ -18,7 +18,6 @@
 
 import abc
 import datetime
-import glob
 import pathlib
 from collections import namedtuple
 import logging
@@ -56,8 +55,6 @@ from google.cloud.aiplatform.compat.services import (
 from google.cloud.aiplatform.compat.types import (
     accelerator_type as gca_accelerator_type,
 )
-
-_logger = logging.getLogger(__name__)
 
 VertexAiServiceClient = TypeVar(
     "VertexAiServiceClient",
@@ -594,123 +591,3 @@ def _timestamped_copy_to_gcs(
 
     gcs_path = "".join(["gs://", "/".join([blob.bucket.name, blob.name])])
     return gcs_path
-
-
-def upload_to_gcs(
-    source_path: str,
-    destination_uri: str,
-    project: Optional[str] = None,
-    credentials: Optional[auth_credentials.Credentials] = None,
-):
-    """Uploads local files to GCS.
-
-    After upload the `destination_uri` will contain the same data as the `source_path`.
-
-    Args:
-        source_path: Required. Path of the local data to copy to GCS.
-        destination_uri: Required. GCS URI where the data should be uploaded.
-        project: Optional. Google Cloud Project that contains the staging bucket.
-        credentials: The custom credentials to use when making API calls.
-            If not provided, default credentials will be used.
-    """
-    source_path_obj = pathlib.Path(source_path)
-    if not source_path_obj.exists():
-        raise RuntimeError(f"Source path does not exist: {source_path}")
-
-    project = project or initializer.global_config.project
-    credentials = credentials or initializer.global_config.credentials
-
-    storage_client = storage.Client(project=project, credentials=credentials)
-    if source_path_obj.is_dir():
-        source_file_paths = glob.glob(
-            pathname=str(source_path_obj / "**"), recursive=True
-        )
-        for source_file_path in source_file_paths:
-            source_file_path_obj = pathlib.Path(source_file_path)
-            if source_file_path_obj.is_dir():
-                continue
-            source_file_relative_path_obj = source_file_path_obj.relative_to(
-                source_path_obj
-            )
-            source_file_relative_posix_path = source_file_relative_path_obj.as_posix()
-            destination_file_uri = (
-                destination_uri.rstrip("/") + "/" + source_file_relative_posix_path
-            )
-            _logger.debug(f'Uploading "{source_file_path}" to "{destination_file_uri}"')
-            destination_blob = storage.Blob.from_string(
-                destination_file_uri, client=storage_client
-            )
-            destination_blob.upload_from_filename(filename=source_file_path)
-    else:
-        source_file_path = source_path
-        destination_file_uri = destination_uri
-        _logger.debug(f'Uploading "{source_file_path}" to "{destination_file_uri}"')
-        destination_blob = storage.Blob.from_string(
-            destination_file_uri, client=storage_client
-        )
-        destination_blob.upload_from_filename(filename=source_file_path)
-
-
-def stage_local_data_in_gcs(
-    data_path: str,
-    staging_gcs_dir: Optional[str] = None,
-    project: Optional[str] = None,
-    location: Optional[str] = None,
-    credentials: Optional[auth_credentials.Credentials] = None,
-) -> str:
-    """Stages a local data in GCS.
-
-    The file copied to GCS is the name of the local file prepended with an
-    "aiplatform-{timestamp}-" string.
-
-    Args:
-        data_path: Required. Path of the local data to copy to GCS.
-        staging_gcs_dir:
-            Optional. Google Cloud Storage bucket to be used for data staging.
-        project: Optional. Google Cloud Project that contains the staging bucket.
-        location: Optional. Google Cloud location to use for the staging bucket.
-        credentials: The custom credentials to use when making API calls.
-            If not provided, default credentials will be used.
-
-    Returns:
-        Google Cloud Storage URI of the staged data.
-    """
-    data_path_obj = pathlib.Path(data_path)
-
-    if not data_path_obj.exists():
-        raise RuntimeError(f"Local data does not exist: data_path='{data_path}'")
-
-    staging_gcs_dir = staging_gcs_dir or initializer.global_config.staging_bucket
-    if not staging_gcs_dir:
-        project = project or initializer.global_config.project
-        location = location or initializer.global_config.location
-        # Creating the bucket if it does not exist.
-        # Currently we only do this when staging_gcs_dir is not specified.
-        staging_bucket_name = project + "-staging"
-        client = storage.Client(project=project, credentials=credentials)
-        staging_bucket = storage.Bucket(client=client, name=staging_bucket_name)
-        if not staging_bucket.exists():
-            _logger.info(f'Creating staging GCS bucket "{staging_bucket_name}"')
-            staging_bucket = client.create_bucket(
-                bucket_or_name=staging_bucket, project=project, location=location,
-            )
-        staging_gcs_dir = "gs://" + staging_bucket_name
-
-    timestamp = datetime.datetime.now().isoformat(sep="-", timespec="milliseconds")
-    staging_gcs_subdir = (
-        staging_gcs_dir.rstrip("/") + "/vertex_ai_auto_staging/" + timestamp
-    )
-
-    staged_data_uri = staging_gcs_subdir
-    if data_path_obj.is_file():
-        staged_data_uri = staging_gcs_subdir + "/" + data_path_obj.name
-
-    _logger.info(f'Uploading "{data_path}" to "{staged_data_uri}"')
-    upload_to_gcs(
-        source_path=data_path,
-        destination_uri=staged_data_uri,
-        project=project,
-        credentials=credentials,
-    )
-
-    return staged_data_uri
